Added Timer Class and basic interfacing
New Timer class to track the time the game has been running for. Will be used to end the game if the time runs out, and a 'bonus time' pick up is to be added at a later point.

Added interfacing for Timer class to State and Level, and timer actors into the test levels.