Added drawTimer() function
added function to draw the Timer onto the canvas. Currently just drawn in a static position (top left corner of the canvas), making the position of the actor in the map irrelevant - it just needs to be present somewhere in the level to be drawn.
If colliding with the timer is to do anything, this will need to be altered so the object moves with the screen.