bugfix with isAbove
In a level that contained multiple of the actor type being searched for, the location of the return false statement would cause only the first actor in the state.actors array to be checked, instead of iterating through all of them.

exit clause was moved to the correct location, bug fixed.